[CWS] extract remote storage to separate package
